+++ core/dbt/include/global_project/macros/materializations/helpers.sql
@@ -63,3 +63,12 @@
     identifier=relation.identifier
   )) -%}
 {% endmacro %}
+
+
+{% macro should_full_refresh() %}
+  {% set config_full_refresh = config.get('full_refresh', false) %}
+  {% if config_full_refresh is none %}
+    {% set config_full_refresh = flags.FULL_REFRESH %}
+  {% endif %}
+  {% do return(config_full_refresh) %}
+{% endmacro %}